feat: add init/build/deploy commands and customizable namespace
This commit comprises some fairly large changes in the codebase.
The 'create' command has been extracted into 'init', 'bulid' and
'deploy' commands. The 'create' command remains, but now delegates
most of its work to these other three. This also has resulted in
some rework of the various flags. One example is the removal of
the '--local' flag from create, since this can be achieved now by
just running 'faas init'.

There are still a number of changes needed for this to land.

* I've removed some of the 'internal' stub code
* The environment variables have been removed (for the most part)
  from --help text and should be returned
* The progress meter is a little wonky now
* The service address is incorrect when printed
* There are no tests, and I'm pretty sure I've broken what
  tests already existed

It's a start - a work in progress.