src: move FunctionConfigForBuild to faas.FunctionConfiguration
Based on review feedback, this function really should not be part
of the CLI itself, but rather in the client library. In this case,
I have moved it to function.go as part of the faas package.

Currently it provides the ability to override the image tag, but
probably should be modified to also allow for a change to the function
name as well.