chore(ci): fix release binary uploads
There was a typo in the upload part of the CI. Also, there was a section
that (thankfully) did not run, which would have created a second release.
Moved the release-please action later in CI so less time elapses between
the creation of the release, and the upload of the binaries.

Signed-off-by: Lance Ball <lball@redhat.com>