feat: add health probes to node & quarkus service
Until the next release of Quarkus in November, we'll need to have some special
checks so that we don't add these probes to Quarkus based functions. In the
meantime, we can at least set the liveness and readiness endpoints on the
Node.js/Go functions to /health/liveness and /health/readiness respectively.

Fixes: https://github.com/boson-project/faas/issues/169

Signed-off-by: Lance Ball <lball@redhat.com>