chore: pull and run pkger on main branch CI
I started this by trying to run a build matrix, but it got complicated pretty
quickly, since this build actually does `make cross-platform`. If we want to
also run a platform matrix when things land on `main`, I think it would be best
if we modified `pull_requests.yaml` to also run on `main` as well as on pull
requests (and be renamed).

Signed-off-by: Lance Ball <lball@redhat.com>