docs: major overhaul of docs (#284)
This commit makes a fairly substantial change to the docs directory.
My intent was to simplify it as much as possible, in order to make
finding relevant content easier. In almost all cases where there was
a stub document with little to no content, that document has been
removed. I have grouped the developer, cli, language and integration
docs under docs/guides and moved the provisioning documents to their
own directory.

I have also added the Node.js language/template/api reference that was
originally written for boson-project/functions. This repo may be a better
place for this and maybe all documentation.

Signed-off-by: Lance Ball <lball@redhat.com>