feat: builders/buildpacks configured in client
We have provided support for modifiying the builders and buildpacks
in func.yaml but until now there has not been a way to set these values
when creating a new function with `client.Create(cfg)`. Any values
provided in `cfg` for the builders and buildpacks were ignored and the
builtin defaults were used.

Signed-off-by: Lance Ball <lball@redhat.com>