feat: periodically update progress during build
Some feedback from the user interviews indicated that people sometimes
get confused by how long it takes to  build a function the first time
and are tempted to kill the process. This change updates the progress
message every five seconds for twenty seconds, and if a build takes longer
than that the message remains "Still building".

Not sure if this is really needed, but it was a fun and easy fix.

Signed-off-by: Lance Ball <lball@redhat.com>